Implement a first version of the sliding window visualization

+++ mining/TopicModellingResult.cs
@@ -3,6 +3,6 @@ using System.Collections.Immutable;
 namespace Ritgard.Mining;
 
 public record TopicModellingResult(
-    ImmutableDictionary<long, Topic> Topics,
+    ImmutableDictionary<int, Topic> Topics,
     ImmutableDictionary<long, TopicItem> Items
 );
